requirements.txt
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
Flask-Migrate
Flask-Restful
passlib
marshmallow
python-dotenv
gunicorn
mysqlclient